requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-filter==1.1.0
django==2.0.2
djangorestframework==3.7.7
markdown==2.6.11
psycopg2-binary==2.7.4
pytz==2018.3              # via django
